Revert "expand overworlds to u16"
This reverts commit deb624f0be1ed5d00399eebe6760b023e6da5e84.

+++ include/union_room.h
@@ -96,7 +96,7 @@ struct WirelessLink_Group
 struct UnionRoomObject
 {
     u8 state;
-    u16 gfxId;
+    u8 gfxId;
     s8 animState;
     u8 schedAnim;
 };
